requirements.txt
python-telegram-bot==20.7
httpx==0.26.0
aiofiles==23.2.1
python-dotenv==1.0.0
pydantic==2.5.3
tenacity==8.2.3
loguru==0.7.2
